Backend/requirements.txt
asgiref==3.8.1
beautifulsoup4==4.13.3
crispy-bootstrap4==2024.10
crispy-bootstrap5==2024.10
diff-match-patch==20241021
Django==4.2.19
django-bootstrap-v5==1.0.11
django-bootstrap4==24.4
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-filter==25.1
django-import-export==4.3.7
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
et_xmlfile==2.0.0
openpyxl==3.1.5
pillow==11.1.0
PyJWT==2.9.0
PyMuPDF==1.25.3
soupsieve==2.6
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.12.2
